+++ lib/diaspora/mentionable.rb
@@ -1,23 +1,20 @@
 
 module Diaspora::Mentionable
 
-  # regex for finding mention markup in plain text
-  # ex.
+  # regex for finding mention markup in plain text:
+  #   "message @{user@pod.net} text"
+  # it can also contain a name, which gets used as the link text:
   #   "message @{User Name; user@pod.net} text"
   #   will yield "User Name" and "user@pod.net"
-  REGEX = /(@\{(.+?; [^\}]+)\})/
+  REGEX = /@\{(?:([^\}]+?); )?([^\} ]+)\}/
 
   # class attribute that will be added to all mention html links
   PERSON_HREF_CLASS = "mention hovercardable"
 
   def self.mention_attrs(mention_str)
-    mention = mention_str.match(REGEX)[2]
-    del_pos = mention.rindex(/;/)
+    name, diaspora_id = mention_str.match(REGEX).captures
 
-    name = mention[0..(del_pos - 1)].strip
-    handle = mention[(del_pos + 1)..-1].strip
-
-    [name, handle]
+    [name.try(:strip), diaspora_id.strip]
   end
 
   # takes a message text and returns the text with mentions in (html escaped)
@@ -32,8 +29,8 @@ module Diaspora::Mentionable
     people = [*people]
 
     msg_text.to_s.gsub(REGEX) {|match_str|
-      name, handle = mention_attrs(match_str)
-      person = people.find {|p| p.diaspora_handle == handle }
+      name, diaspora_id = mention_attrs(match_str)
+      person = people.find {|p| p.diaspora_handle == diaspora_id }
 
       ERB::Util.h(MentionsInternal.mention_link(person, name, opts))
     }
@@ -45,10 +42,7 @@ module Diaspora::Mentionable
   # @param [String] text containing mentions
   # @return [Array<Person>] array of people
   def self.people_from_string(msg_text)
-    identifiers = msg_text.to_s.scan(REGEX).map do |match_str|
-      _, identifier = mention_attrs(match_str.first)
-      identifier if Validation::Rule::DiasporaId.new.valid_value?(identifier)
-    end
+    identifiers = msg_text.to_s.scan(REGEX).map {|match_str| match_str.second.strip }
 
     identifiers.compact.uniq.map {|identifier| find_or_fetch_person_by_identifier(identifier) }.compact
   end
@@ -64,18 +58,30 @@ module Diaspora::Mentionable
     mentioned_ppl = people_from_string(msg_text)
 
     msg_text.to_s.gsub(REGEX) {|match_str|
-      name, handle = mention_attrs(match_str)
-      person = mentioned_ppl.find {|p| p.diaspora_handle == handle }
+      name, diaspora_id = mention_attrs(match_str)
+      person = mentioned_ppl.find {|p| p.diaspora_handle == diaspora_id }
       mention = MentionsInternal.profile_link(person, name) unless allowed_people.include?(person.id)
 
       mention || match_str
     }
   end
 
-  private
+  # Regex to find mentions with new syntax, only used for backporting to old syntax
+  NEW_SYNTAX_REGEX = /@\{[^ ]+\}/
+
+  # replaces new syntax with old syntax, to be compatible with old pods
+  # @deprecated remove when most of the posts can handle the new syntax
+  def self.backport_mention_syntax(text)
+    text.to_s.gsub(NEW_SYNTAX_REGEX) do |match_str|
+      _, diaspora_id = mention_attrs(match_str)
+      person = find_or_fetch_person_by_identifier(diaspora_id)
+      old_syntax = "@{#{person.name}; #{diaspora_id}}" if person
+      old_syntax || match_str
+    end
+  end
 
   private_class_method def self.find_or_fetch_person_by_identifier(identifier)
-    Person.find_or_fetch_by_identifier(identifier)
+    Person.find_or_fetch_by_identifier(identifier) if Validation::Rule::DiasporaId.new.valid_value?(identifier)
   rescue DiasporaFederation::Discovery::DiscoveryError
     nil
   end
@@ -113,5 +119,4 @@ module Diaspora::Mentionable
       "[#{display_name.presence || person.name}](#{local_or_remote_person_path(person)})"
     end
   end
-
 end